requirements.txt
ollama
python-docx
pdfplumber
pyttsx3